feat(landing): replace auth-router root with public landing + metadata
/ now serves the marketing landing page. Authenticated visitors still
redirect to /dashboard via useAuth/useRouter; setup-needed users still
get sent to /setup.

Split pattern:
- app/page.tsx — server component, exports Metadata + OpenGraph.
- components/landing/LandingPageBody.tsx — client component handling
  the auth redirect and composing the 5 landing sections.

The split is required because Next.js App Router only accepts metadata
exports from server components. This is the minimum metadata for
social-share unfurls and browser tabs; full SEO baseline (sitemap,
robots, JSON-LD, Lighthouse audit) lands in roadmap L5.2.
