+++ frontend/app/page.tsx
@@ -1,28 +1,19 @@
-"use client";
+import type { Metadata } from "next";
+import LandingPageBody from "@/components/landing/LandingPageBody";
 
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { useAuth } from "@/components/auth/AuthProvider";
+export const metadata: Metadata = {
+  title: "The Better Decision — know your money, plan what's next",
+  description:
+    "A finance app for normal people. Know what you have, what's coming, and where it goes — without the spreadsheet fatigue.",
+  openGraph: {
+    title: "The Better Decision — know your money, plan what's next",
+    description:
+      "A finance app for normal people. Know what you have, what's coming, and where it goes.",
+    type: "website",
+    siteName: "The Better Decision",
+  },
+};
 
-export default function RootPage() {
-  const { user, loading, needsSetup } = useAuth();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (!loading) {
-      if (needsSetup) {
-        router.replace("/setup");
-      } else if (user) {
-        router.replace("/dashboard");
-      } else {
-        router.replace("/login");
-      }
-    }
-  }, [user, loading, needsSetup, router]);
-
-  return (
-    <div className="flex h-screen items-center justify-center bg-bg">
-      <div className="h-6 w-6 animate-spin rounded-full border-2 border-border border-t-accent" />
-    </div>
-  );
+export default function LandingPage() {
+  return <LandingPageBody />;
 }
